feature: Add #EXT-X-VERSION tag
The EXT-X-VERSION tag indicates the compatibility version of the
Playlist file, its associated media, and its server.
The EXT-X-VERSION tag applies to the entire Playlist file.  Its
format is: #EXT-X-VERSION:<n>
where n is an integer indicating the protocol compatibility version
number.

It MUST appear in all Playlists containing tags or attributes that
are not compatible with protocol version 1 to support
interoperability with older clients.

A Playlist file MUST NOT contain more than one EXT-X-VERSION tag.  If
a client encounters a Playlist with multiple EXT-X-VERSION tags, it
MUST reject it.

Currently hardcoded to version 3 until the need to specify it emerge.