feature: Add EXT-X-STREAM-INF
EXT-X-STREAM-INF
https://tools.ietf.org/html/draft-pantos-http-live-streaming-18#section-4.3.3

The EXT-X-STREAM-INF tag specifies a Variant Stream, which is a set of Renditions
which can be combined to play the presentation. The attributes of the tag provide
information about the Variant Stream. The URI line that follows the EXT-X-STREAM-INF
tag specifies a Media Playlist that carries a Rendition of the Variant
Stream. The URI line is REQUIRED.  Clients that do not support multiple video renditions
SHOULD play this Rendition. Its format is:
 #EXT-X-STREAM-INF:<attribute-list>
 <URI>

Currently only URI implemented, follwing attributes on demand.