feature: Add bandwidth to EXT-X-STREAM-INF tag
The bandwidth is a required param to the EXT-X-STREAM-INF tag

The value is a decimal-integer of bits per second. It represents the peak segment bit rate
of the Variant Stream. If all the Media Segments in a Variant Stream have already been created, the BANDWIDTH value MUST
be the largest sum of peak segment bit rates that is produced by any playable combination of Renditions.
An inaccurate value can cause playback stalls or prevent clients from playing the variant.
If the Master Playlist is to be made available before all Media Segments in the presentation have been encoded,
the BANDWIDTH value SHOULD be the BANDWIDTH value of a representative period of similar content, encoded using the
same settings. Every EXT-X-STREAM-INF tag MUST include the BANDWIDTH attribute.